+++ src/test/body_send.rs
@@ -10,7 +10,7 @@ fn content_length_on_str() {
     let resp = post("test://host/content_length_on_str")
         .send_string("Hello World!!!")
         .unwrap();
-    let vec = resp.to_write_vec();
+    let vec = resp.as_write_vec();
     let s = String::from_utf8_lossy(&vec);
     assert!(s.contains("\r\nContent-Length: 14\r\n"));
 }
@@ -24,7 +24,7 @@ fn user_set_content_length_on_str() {
         .set("Content-Length", "12345")
         .send_string("Hello World!!!")
         .unwrap();
-    let vec = resp.to_write_vec();
+    let vec = resp.as_write_vec();
     let s = String::from_utf8_lossy(&vec);
     assert!(s.contains("\r\nContent-Length: 12345\r\n"));
 }
@@ -35,15 +35,15 @@ fn content_length_on_json() {
     test::set_handler("/content_length_on_json", |_unit| {
         test::make_response(200, "OK", vec![], vec![])
     });
-    let mut json = SerdeMap::new();
+    let mut json = serde_json::Map::new();
     json.insert(
         "Hello".to_string(),
-        SerdeValue::String("World!!!".to_string()),
+        serde_json::Value::String("World!!!".to_string()),
     );
     let resp = post("test://host/content_length_on_json")
-        .send_json(SerdeValue::Object(json))
+        .send_json(serde_json::Value::Object(json))
         .unwrap();
-    let vec = resp.to_write_vec();
+    let vec = resp.as_write_vec();
     let s = String::from_utf8_lossy(&vec);
     assert!(s.contains("\r\nContent-Length: 20\r\n"));
 }
@@ -57,7 +57,7 @@ fn content_length_and_chunked() {
         .set("Transfer-Encoding", "chunked")
         .send_string("Hello World!!!")
         .unwrap();
-    let vec = resp.to_write_vec();
+    let vec = resp.as_write_vec();
     let s = String::from_utf8_lossy(&vec);
     assert!(s.contains("Transfer-Encoding: chunked\r\n"));
     assert!(!s.contains("\r\nContent-Length:\r\n"));
@@ -73,7 +73,7 @@ fn str_with_encoding() {
         .set("Content-Type", "text/plain; charset=iso-8859-1")
         .send_string("Hällo Wörld!!!")
         .unwrap();
-    let vec = resp.to_write_vec();
+    let vec = resp.as_write_vec();
     assert_eq!(
         &vec[vec.len() - 14..],
         //H  ä    l    l    o    _   W   ö    r    l    d    !   !   !
@@ -87,15 +87,15 @@ fn content_type_on_json() {
     test::set_handler("/content_type_on_json", |_unit| {
         test::make_response(200, "OK", vec![], vec![])
     });
-    let mut json = SerdeMap::new();
+    let mut json = serde_json::Map::new();
     json.insert(
         "Hello".to_string(),
-        SerdeValue::String("World!!!".to_string()),
+        serde_json::Value::String("World!!!".to_string()),
     );
     let resp = post("test://host/content_type_on_json")
-        .send_json(SerdeValue::Object(json))
+        .send_json(serde_json::Value::Object(json))
         .unwrap();
-    let vec = resp.to_write_vec();
+    let vec = resp.as_write_vec();
     let s = String::from_utf8_lossy(&vec);
     assert!(s.contains("\r\nContent-Type: application/json\r\n"));
 }
@@ -106,16 +106,16 @@ fn content_type_not_overriden_on_json() {
     test::set_handler("/content_type_not_overriden_on_json", |_unit| {
         test::make_response(200, "OK", vec![], vec![])
     });
-    let mut json = SerdeMap::new();
+    let mut json = serde_json::Map::new();
     json.insert(
         "Hello".to_string(),
-        SerdeValue::String("World!!!".to_string()),
+        serde_json::Value::String("World!!!".to_string()),
     );
     let resp = post("test://host/content_type_not_overriden_on_json")
         .set("content-type", "text/plain")
-        .send_json(SerdeValue::Object(json))
+        .send_json(serde_json::Value::Object(json))
         .unwrap();
-    let vec = resp.to_write_vec();
+    let vec = resp.as_write_vec();
     let s = String::from_utf8_lossy(&vec);
     assert!(s.contains("\r\ncontent-type: text/plain\r\n"));
 }