+++ src/unit.rs
@@ -1,4 +1,6 @@
+use std::fmt::{self, Display};
 use std::io::{self, Write};
+use std::ops::Range;
 use std::time;
 
 use log::debug;
@@ -7,6 +9,7 @@ use url::Url;
 #[cfg(feature = "cookies")]
 use cookie::Cookie;
 
+use crate::agent::RedirectAuthHeaders;
 use crate::body::{self, BodySize, Payload, SizedReader};
 use crate::error::{Error, ErrorKind};
 use crate::header;
@@ -36,13 +39,13 @@ impl Unit {
         agent: &Agent,
         method: &str,
         url: &Url,
-        headers: &[Header],
+        mut headers: Vec<Header>,
         body: &SizedReader,
         deadline: Option<time::Instant>,
     ) -> Self {
         //
 
-        let (is_transfer_encoding_set, mut is_chunked) = get_header(headers, "transfer-encoding")
+        let (is_transfer_encoding_set, mut is_chunked) = get_header(&headers, "transfer-encoding")
             // if the user has set an encoding header, obey that.
             .map(|enc| {
                 let is_transfer_encoding_set = !enc.is_empty();
@@ -55,12 +58,12 @@ impl Unit {
             // otherwise, no chunking.
             .unwrap_or((false, false));
 
-        let extra_headers = {
+        let mut extra_headers = {
             let mut extra = vec![];
 
             // chunking and Content-Length headers are mutually exclusive
             // also don't write this if the user has set it themselves
-            if !is_chunked && get_header(headers, "content-length").is_none() {
+            if !is_chunked && get_header(&headers, "content-length").is_none() {
                 // if the payload is of known size (everything beside an unsized reader), set
                 // Content-Length,
                 // otherwise, use the chunked Transfer-Encoding (only if no other Transfer-Encoding
@@ -82,7 +85,7 @@ impl Unit {
             let username = url.username();
             let password = url.password().unwrap_or("");
             if (!username.is_empty() || !password.is_empty())
-                && get_header(headers, "authorization").is_none()
+                && get_header(&headers, "authorization").is_none()
             {
                 let encoded = base64::encode(&format!("{}:{}", username, password));
                 extra.push(Header::new("Authorization", &format!("Basic {}", encoded)));
@@ -94,11 +97,7 @@ impl Unit {
             extra
         };
 
-        let headers: Vec<_> = headers
-            .iter()
-            .chain(extra_headers.iter())
-            .cloned()
-            .collect();
+        headers.append(&mut extra_headers);
 
         Unit {
             agent: agent.clone(),
@@ -202,17 +201,32 @@ pub(crate) fn connect(
             }
             _ => break resp,
         };
+
+        let keep_auth_header = can_propagate_authorization_on_redirect(
+            &unit.agent.config.redirect_auth_headers,
+            url,
+            &new_url,
+        );
+
         debug!("redirect {} {} -> {}", resp.status(), url, new_url);
-        history.push(unit.url.to_string());
+        history.push(unit.url);
         body = Payload::Empty.into_read();
-        unit.headers.retain(|h| h.name() != "Content-Length");
+
+        // reuse the previous header vec on redirects.
+        let mut headers = unit.headers;
+
+        // on redirects we don't want to keep "content-length". we also might want to
+        // strip away "authorization" to ensure credentials are not leaked.
+        headers.retain(|h| {
+            !h.is_name("content-length") && (!h.is_name("authorization") || keep_auth_header)
+        });
 
         // recreate the unit to get a new hostname and cookies for the new host.
         unit = Unit::new(
             &unit.agent,
             &new_method,
             &new_url,
-            &unit.headers,
+            headers,
             &body,
             unit.deadline,
         );
@@ -226,7 +240,7 @@ fn connect_inner(
     unit: &Unit,
     use_pooled: bool,
     body: SizedReader,
-    previous: &[String],
+    history: &[Url],
 ) -> Result<Response, Error> {
     let host = unit
         .url
@@ -244,7 +258,7 @@ fn connect_inner(
         debug!("sending request {} {}", method, url);
     }
 
-    let send_result = send_prelude(unit, &mut stream, !previous.is_empty());
+    let send_result = send_prelude(unit, &mut stream);
 
     if let Err(err) = send_result {
         if is_recycled {
@@ -252,7 +266,7 @@ fn connect_inner(
             // we try open a new connection, this time there will be
             // no connection in the pool. don't use it.
             // NOTE: this recurses at most once because `use_pooled` is `false`.
-            return connect_inner(unit, false, body, previous);
+            return connect_inner(unit, false, body, history);
         } else {
             // not a pooled connection, propagate the error.
             return Err(err.into());
@@ -281,7 +295,7 @@ fn connect_inner(
             debug!("retrying request {} {}: {}", method, url, err);
             let empty = Payload::Empty.into_read();
             // NOTE: this recurses at most once because `use_pooled` is `false`.
-            return connect_inner(unit, false, empty, previous);
+            return connect_inner(unit, false, empty, history);
         }
         Err(e) => return Err(e),
         Ok(resp) => resp,
@@ -351,19 +365,42 @@ fn connect_socket(unit: &Unit, hostname: &str, use_pooled: bool) -> Result<(Stre
     Ok((stream?, false))
 }
 
+fn can_propagate_authorization_on_redirect(
+    redirect_auth_headers: &RedirectAuthHeaders,
+    prev_url: &Url,
+    url: &Url,
+) -> bool {
+    fn scheme_is_https(url: &Url) -> bool {
+        url.scheme() == "https" || (cfg!(test) && url.scheme() == "test")
+    }
+
+    match redirect_auth_headers {
+        RedirectAuthHeaders::Never => false,
+        RedirectAuthHeaders::SameHost => {
+            let host = url.host_str();
+            let is_https = scheme_is_https(url);
+
+            let prev_host = prev_url.host_str();
+            let prev_is_https = scheme_is_https(prev_url);
+
+            let same_scheme_or_more_secure =
+                is_https == prev_is_https || (!prev_is_https && is_https);
+
+            host == prev_host && same_scheme_or_more_secure
+        }
+    }
+}
+
 /// Send request line + headers (all up until the body).
 #[allow(clippy::write_with_newline)]
-fn send_prelude(unit: &Unit, stream: &mut Stream, redir: bool) -> io::Result<()> {
+fn send_prelude(unit: &Unit, stream: &mut Stream) -> io::Result<()> {
     // build into a buffer and send in one go.
-    let mut prelude: Vec<u8> = vec![];
+    let mut prelude = PreludeBuilder::new();
 
     // request line
-    write!(
-        prelude,
-        "{} {}{}{} HTTP/1.1\r\n",
-        unit.method,
+    prelude.write_request_line(
+        &unit.method,
         unit.url.path(),
-        if unit.url.query().is_some() { "?" } else { "" },
         unit.url.query().unwrap_or_default(),
     )?;
 
@@ -378,40 +415,113 @@ fn send_prelude(unit: &Unit, stream: &mut Stream, redir: bool) -> io::Result<()>
                     _ => 0,
                 };
                 if scheme_default != 0 && scheme_default == port {
-                    write!(prelude, "Host: {}\r\n", host)?;
+                    prelude.write_header("Host", host)?;
                 } else {
-                    write!(prelude, "Host: {}:{}\r\n", host, port)?;
+                    prelude.write_header("Host", format_args!("{}:{}", host, port))?;
                 }
             }
             None => {
-                write!(prelude, "Host: {}\r\n", host)?;
+                prelude.write_header("Host", host)?;
             }
         }
     }
     if !header::has_header(&unit.headers, "user-agent") {
-        write!(prelude, "User-Agent: {}\r\n", &unit.agent.config.user_agent)?;
+        prelude.write_header("User-Agent", &unit.agent.config.user_agent)?;
     }
     if !header::has_header(&unit.headers, "accept") {
-        write!(prelude, "Accept: */*\r\n")?;
+        prelude.write_header("Accept", "*/*")?;
     }
 
     // other headers
     for header in &unit.headers {
-        if !redir || !header.is_name("Authorization") {
-            if let Some(v) = header.value() {
-                write!(prelude, "{}: {}\r\n", header.name(), v)?;
+        if let Some(v) = header.value() {
+            if is_header_sensitive(header) {
+                prelude.write_sensitive_header(header.name(), v)?;
+            } else {
+                prelude.write_header(header.name(), v)?;
             }
         }
     }
 
     // finish
-    write!(prelude, "\r\n")?;
+    prelude.finish()?;
 
-    debug!("writing prelude: {}", String::from_utf8_lossy(&prelude));
+    debug!("writing prelude: {}", prelude);
     // write all to the wire
-    stream.write_all(&prelude[..])?;
+    stream.write_all(prelude.as_slice())?;
 
     Ok(())
+}
+
+fn is_header_sensitive(header: &Header) -> bool {
+    header.is_name("Authorization") || header.is_name("Cookie")
+}
+
+struct PreludeBuilder {
+    prelude: Vec<u8>,
+    // Sensitive information to be omitted in debug logging
+    sensitive_spans: Vec<Range<usize>>,
+}
+
+impl PreludeBuilder {
+    fn new() -> Self {
+        PreludeBuilder {
+            prelude: Vec::with_capacity(256),
+            sensitive_spans: Vec::new(),
+        }
+    }
+
+    fn write_request_line(&mut self, method: &str, path: &str, query: &str) -> io::Result<()> {
+        write!(self.prelude, "{} {}", method, path,)?;
+        if !query.is_empty() {
+            write!(self.prelude, "?{}", query)?;
+        }
+        write!(self.prelude, " HTTP/1.1\r\n")?;
+        Ok(())
+    }
+
+    fn write_header(&mut self, name: &str, value: impl Display) -> io::Result<()> {
+        write!(self.prelude, "{}: {}\r\n", name, value)
+    }
+
+    fn write_sensitive_header(&mut self, name: &str, value: impl Display) -> io::Result<()> {
+        write!(self.prelude, "{}: ", name)?;
+        let start = self.prelude.len();
+        write!(self.prelude, "{}", value)?;
+        let end = self.prelude.len();
+        self.sensitive_spans.push(start..end);
+        write!(self.prelude, "\r\n")?;
+        Ok(())
+    }
+
+    fn finish(&mut self) -> io::Result<()> {
+        write!(self.prelude, "\r\n")
+    }
+
+    fn as_slice(&self) -> &[u8] {
+        &self.prelude
+    }
+}
+
+impl fmt::Display for PreludeBuilder {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        let mut pos = 0;
+        for span in &self.sensitive_spans {
+            write!(
+                f,
+                "{}",
+                String::from_utf8_lossy(&self.prelude[pos..span.start])
+            )?;
+            write!(f, "***")?;
+            pos = span.end;
+        }
+        write!(
+            f,
+            "{}",
+            String::from_utf8_lossy(&self.prelude[pos..]).trim_end()
+        )?;
+        Ok(())
+    }
 }
 
 /// Investigate a response for "Set-Cookie" headers.