+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# 2.3.2
+ * Fix regression in 2.3.x for feature native-certs (#441)
+
+# 2.3.1
+ * Don't panic when given an invalid DNS name (#436).
+ * Update to rustls-native-certs v0.6 (#432).
+
+# 2.3.0
+ * Upgrade to rustls 0.20 (#427).
+ * Make test mocks of Response more accurate by removing newline (#423).
+ * Redact sensitive headers when logging prelude (#414).
+
 # 2.2.0
  * Update to latest dependencies
  * Add SOCKS4 support (#410).